move funtions into another file
